migrate tests for package_activity_list
