+++ ckanext/unhcr/tests/nose/test_actions.py
@@ -181,102 +181,6 @@ class TestPrivateResources(base.FunctionalTestBase):
         assert_equals(res.status_int, 200)
 
 
-class TestPackageActivityList(base.FunctionalTestBase):
-
-    def setup(self):
-        super(TestPackageActivityList, self).setup()
-
-        self.sysadmin = core_factories.Sysadmin(name='sysadmin', id='sysadmin')
-        self.container1_admin = core_factories.User()
-        self.container1_member = core_factories.User()
-        self.container1 = factories.DataContainer(
-            users=[
-                {"name": self.container1_admin["name"], "capacity": "admin"},
-                {"name": self.container1_member["name"], "capacity": "member"},
-            ]
-        )
-        self.dataset1 = factories.Dataset(
-            owner_org=self.container1["id"], visibility="private"
-        )
-        self.resource1 = factories.Resource(
-            package_id=self.dataset1['id'],
-            upload=mocks.FakeFileStorage(),
-            url = "http://fakeurl/test.txt",
-            url_type='upload',
-        )
-        helpers.create_curation_activity(
-            'dataset_approved',
-            self.dataset1['id'],
-            self.dataset1['name'],
-            self.sysadmin['id'],
-            message='asdf'
-        )
-        log_download_activity({'user': self.sysadmin['name']}, self.resource1['id'])
-
-    def test_container_admin(self):
-        context = {
-            'user': self.container1_admin['name'],
-            'package': model.package.Package.get(self.dataset1['id'])
-        }
-        data_dict = {
-            'id': self.dataset1['id'],
-            'get_internal_activities': True
-        }
-        activities = toolkit.get_action('package_activity_list')(context, data_dict)
-        # a container admin can see all the internal activities
-        assert_equals(2, len(activities))
-        assert_equals('download resource', activities[0]['activity_type'])
-        assert_equals('dataset_approved', activities[1]['data']['curation_activity'])
-
-    def test_dataset_editor(self):
-        collaborator = core_factories.User()
-        core_helpers.call_action(
-            'dataset_collaborator_create',
-            id=self.dataset1['id'],
-            user_id=collaborator['id'],
-            capacity='editor',
-        )
-        context = {
-            'user': collaborator['name'],
-            'package': model.package.Package.get(self.dataset1['id'])
-        }
-        data_dict = {
-            'id': self.dataset1['id'],
-            'get_internal_activities': True
-        }
-        activities = toolkit.get_action('package_activity_list')(context, data_dict)
-        # a dataset editor can only see the curation activities
-        assert_equals(1, len(activities))
-        assert_equals('dataset_approved', activities[0]['data']['curation_activity'])
-
-    def test_container_member(self):
-        context = {
-            'user': self.container1_member['name'],
-            'package': model.package.Package.get(self.dataset1['id'])
-        }
-        data_dict = {
-            'id': self.dataset1['id'],
-            'get_internal_activities': True
-        }
-        action = toolkit.get_action('package_activity_list')
-        # a container member can't see any internal activities
-        assert_raises(toolkit.NotAuthorized, action, context, data_dict)
-
-    def test_unprivileged_user(self):
-        normal_user = core_factories.User()
-        context = {
-            'user': normal_user['name'],
-            'package': model.package.Package.get(self.dataset1['id'])
-        }
-        data_dict = {
-            'id': self.dataset1['id'],
-            'get_internal_activities': True
-        }
-        action = toolkit.get_action('package_activity_list')
-        # an unprivileged user can't see any internal activities
-        assert_raises(toolkit.NotAuthorized, action, context, data_dict)
-
-
 class TestDatastoreAuthRestrictedDownloads(base.FunctionalTestBase):
 
     def setup(self):
